# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: PoupaMais - Personal Finance API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /categories:
    get:
      operationId: getCategories
      tags: [categories]
      summary: List all categories
      responses:
        "200":
          description: List of categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
    post:
      operationId: createCategory
      tags: [categories]
      summary: Create a custom category
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryInput"
      responses:
        "201":
          description: Category created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
  /categories/{id}:
    patch:
      operationId: updateCategory
      tags: [categories]
      summary: Update a category
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
    delete:
      operationId: deleteCategory
      tags: [categories]
      summary: Delete a category
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
  /cards:
    get:
      operationId: getCards
      tags: [cards]
      summary: List all credit cards
      responses:
        "200":
          description: List of cards
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Card"
    post:
      operationId: createCard
      tags: [cards]
      summary: Add a credit card
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CardInput"
      responses:
        "201":
          description: Card added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Card"
  /cards/{id}:
    patch:
      operationId: updateCard
      tags: [cards]
      summary: Update a credit card
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CardUpdate"
      responses:
        "200":
          description: Card updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Card"
    delete:
      operationId: deleteCard
      tags: [cards]
      summary: Delete a credit card
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
  /ai/chat:
    post:
      operationId: chatWithAI
      tags: [ai]
      summary: Send a message to PoupaAI
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AIChatInput"
      responses:
        "200":
          description: AI response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AIChatResponse"
...
  /ai/conversations/{id}/messages:
    get:
      operationId: getAIMessages
      tags: [ai]
      summary: Get messages in a conversation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AIMessage"
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: ["string", "null"]
      required:
        - success
    RegisterInput:
      type: object
      required: [name, email, password]
      properties:
        name:
          type: string
          minLength: 2
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
        currency:
          type: ["string", "null"]
        language:
          type: ["string", "null"]
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
        rememberMe:
          type: ["boolean", "null"]
    AuthResponse:
      type: object
      required: [user, token]
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
    User:
      type: object
      required: [id, name, email, currency, language, isPremium, onboardingCompleted, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        avatarUrl:
          type: ["string", "null"]
        currency:
          type: string
        language:
          type: string
        isPremium:
          type: boolean
        premiumExpiresAt:
          type: ["string", "null"]
        onboardingCompleted:
          type: boolean
        createdAt:
          type: string
...
    Category:
      type: object
      required: [id, name, icon, color, type, isDefault]
      properties:
        id:
          type: integer
        name:
          type: string
        icon:
          type: string
        color:
          type: string
        type:
          type: string
          enum: [income, expense, both]
        isDefault:
          type: boolean
        userId:
          type: ["integer", "null"]
    CategoryInput:
      type: object
      required: [name, icon, color, type]
      properties:
        name:
          type: string
          minLength: 1
        icon:
          type: string
        color:
          type: string
        type:
          type: string
          enum: [income, expense, both]
    CategoryUpdate:
      type: object
      properties:
        name:
          type: ["string", "null"]
        icon:
          type: ["string", "null"]
        color:
          type: ["string", "null"]
        type:
          type: ["string", "null"]
...
    Card:
      type: object
      required: [id, userId, name, lastFourDigits, brand, limit, currentBalance, closingDay, dueDay, color, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        name:
          type: string
        lastFourDigits:
          type: string
        brand:
          type: string
          enum: [visa, mastercard, elo, amex, hipercard, other]
        limit:
          type: number
        currentBalance:
          type: number
        closingDay:
          type: integer
        dueDay:
          type: integer
        color:
          type: string
        createdAt:
          type: string
    CardInput:
      type: object
      required: [name, lastFourDigits, brand, limit, closingDay, dueDay]
      properties:
        name:
          type: string
          minLength: 1
        lastFourDigits:
          type: string
          minLength: 4
          maxLength: 4
        brand:
          type: string
          enum: [visa, mastercard, elo, amex, hipercard, other]
        limit:
          type: number
          minimum: 1
        currentBalance:
          type: ["number", "null"]
        closingDay:
          type: integer
          minimum: 1
          maximum: 31
        dueDay:
          type: integer
          minimum: 1
          maximum: 31
        color:
          type: ["string", "null"]
    CardUpdate:
      type: object
      properties:
        name:
          type: ["string", "null"]
        limit:
          type: ["number", "null"]
        currentBalance:
          type: ["number", "null"]
        closingDay:
          type: ["integer", "null"]
        dueDay:
          type: ["integer", "null"]
        color:
          type: ["string", "null"]
...
    AIChatInput:
      type: object
      required: [message]
      properties:
        message:
          type: string
          minLength: 1
        conversationId:
          type: ["integer", "null"]
    AIChatResponse:
      type: object
      required: [message, conversationId]
      properties:
        message:
          type: string
        conversationId:
          type: integer
        suggestions:
          type: array
          items:
            type: string
...
    AIMessage:
      type: object
      required: [id, conversationId, role, content, createdAt]
      properties:
        id:
          type: integer
        conversationId:
          type: integer
        role:
          type: string
          enum: [user, assistant]
        content:
          type: string
        createdAt:
          type: string