# lib/api-zod/src/generated/types/aIChatResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * PoupaMais - Personal Finance API
 * OpenAPI spec version: 0.1.0
 */

export interface AIChatResponse {
  message: string;
  conversationId: number;
  suggestions?: string[];
}