# lib/api-zod/src/generated/types/categoryUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * PoupaMais - Personal Finance API
 * OpenAPI spec version: 0.1.0
 */

export interface CategoryUpdate {
  /** @nullable */
  name?: string | null;
  /** @nullable */
  icon?: string | null;
  /** @nullable */
  color?: string | null;
  /** @nullable */
  type?: string | null;
}