# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * PoupaMais - Personal Finance API
 * OpenAPI spec version: 0.1.0
 */

export interface User {
  id: number;
  name: string;
  email: string;
  /** @nullable */
  avatarUrl?: string | null;
  currency: string;
  language: string;
  isPremium: boolean;
  /** @nullable */
  premiumExpiresAt?: string | null;
  onboardingCompleted: boolean;
  createdAt: string;
}